requirements.txt
Django
psycopg2-binary
djangorestframework
Pillow
django-cors-headers
django_cleanup
django-markdownx
django-environ
